+++ cmd/indexer/main_test.go
@@ -222,7 +222,7 @@ func initConfig(tCfg testConfig, history bool) config.Config {
 			BatchSize:               500,
 			StartIndex:              startBlock,
 			StopIndex:               endBlock,
-			NumParallelReq:          16,
+			RpcConcurrency:          16,
 			LogRange:                10,
 			NewBlockCheckMillis:     1000,
 			CollectTransactions:     []config.TransactionInfo{txInfo},
@@ -263,7 +263,7 @@ func createIndexer(cfg *config.Config, db *gorm.DB) (*core.Engine, error) {
 		return nil, errors.Wrap(err, "Invalid node URL in config")
 	}
 
-	ethClient, err := chain.DialRPCNode(nodeURL, cfg.Chain.ChainType)
+	ethClient, err := chain.DialRPCNode(nodeURL, cfg.Chain.ChainType, cfg.Indexer.RpcConcurrency)
 	if err != nil {
 		return nil, errors.Wrap(err, "Could not connect to the RPC nodes")
 	}